@@ -5,33 +5,22 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.7.17</version>
+		<version>3.5.5</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.flower</groupId>
-	<artifactId>spirit</artifactId>
+	<artifactId>StreamVault</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>spirit</name>
+	<name>StreamVault</name>
 	<description>Demo project for Spring Boot</description>
 	<properties>
-	  <maven.compiler.source>8</maven.compiler.source>
-	  <maven.compiler.target>8</maven.compiler.target>
+	  <maven.compiler.source>17</maven.compiler.source>
+	  <maven.compiler.target>17</maven.compiler.target>
 	</properties>
 	<dependencies>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
-			<exclusions>
-		        <exclusion>
-		            <groupId>com.fasterxml.jackson.core</groupId>
-		            <artifactId>jackson-databind</artifactId>
-		        </exclusion>
-   			 </exclusions>
-		</dependency>
-	    <dependency>
-		    <groupId>com.fasterxml.jackson.core</groupId>
-		    <artifactId>jackson-databind</artifactId>
-		    <version>2.14.2</version><!--$NO-MVN-MAN-VER$-->
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
@@ -53,11 +42,14 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-thymeleaf</artifactId>
 		</dependency>
-        <dependency>
-            <groupId>com.enigmabridge</groupId>
-            <artifactId>hibernate4-sqlite-dialect</artifactId>
-            <version>0.1.2</version>
-        </dependency>
+		
+
+		<dependency>
+		    <groupId>org.hibernate.orm</groupId>
+		    <artifactId>hibernate-community-dialects</artifactId>
+		    <version>6.6.26.Final</version><!--$NO-MVN-MAN-VER$-->
+		</dependency>
+
         <dependency>
             <groupId>org.xerial</groupId>
             <artifactId>sqlite-jdbc</artifactId>
@@ -66,7 +58,7 @@
         <dependency>
 			<groupId>commons-io</groupId>
 			<artifactId>commons-io</artifactId>
-			<version>2.7</version>
+			<version>2.14.0</version>
 		</dependency>
  
         <dependency>
@@ -114,21 +106,6 @@
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
 			</plugin>
-			
-			<plugin>
-	            <groupId>com.spotify</groupId>
-	            <artifactId>docker-maven-plugin</artifactId>
-				<configuration>
-		            <imageName>${docker.image.prefix}/aoonav</imageName>
-		            <dockerDirectory>src/main/docker</dockerDirectory>
-		            <resources>
-		               <resource>
-		                  <directory>${project.build.directory}</directory>
-		                  <include>${project.build.finalName}.jar</include>
-		               </resource>
-		            </resources>
-		         </configuration>
-		       </plugin> 
 		</plugins>
 	</build>
 
