@@ -2,10 +2,10 @@ package com.flower.spirit.dao;
 
 import java.util.List;
 
-import javax.transaction.Transactional;
+import jakarta.transaction.Transactional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 import com.flower.spirit.entity.BiliConfigEntity;
@@ -15,7 +15,7 @@ import com.flower.spirit.entity.BiliConfigEntity;
 
 @Repository
 @Transactional
-public interface BiliConfigDao extends PagingAndSortingRepository<BiliConfigEntity, Integer>, JpaSpecificationExecutor<BiliConfigEntity>{
+public interface BiliConfigDao extends JpaRepository<BiliConfigEntity, Integer>, JpaSpecificationExecutor<BiliConfigEntity>{
 
 	
 	public List<BiliConfigEntity> findAll();
