@@ -1,9 +1,10 @@
 package com.flower.spirit.dao;
 
 import java.util.List;
-import javax.transaction.Transactional;
+import jakarta.transaction.Transactional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 import com.flower.spirit.entity.CollectDataDetailEntity;
 
@@ -12,7 +13,7 @@ import com.flower.spirit.entity.CollectDataDetailEntity;
 
 @Repository
 @Transactional
-public interface CollectdDataDetailDao extends PagingAndSortingRepository<CollectDataDetailEntity, Integer>, JpaSpecificationExecutor<CollectDataDetailEntity>{
+public interface CollectdDataDetailDao extends JpaRepository<CollectDataDetailEntity, Integer>, JpaSpecificationExecutor<CollectDataDetailEntity>{
 
 	
 	public List<CollectDataDetailEntity> findAll();
