@@ -2,16 +2,16 @@ package com.flower.spirit.dao;
 
 import java.util.List;
 
-import javax.transaction.Transactional;
+import jakarta.transaction.Transactional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 import com.flower.spirit.entity.CookiesConfigEntity;
 
 @Repository
 @Transactional
-public interface CookiesConfigDao extends PagingAndSortingRepository<CookiesConfigEntity, Integer>, JpaSpecificationExecutor<CookiesConfigEntity> {
+public interface CookiesConfigDao extends JpaRepository<CookiesConfigEntity, Integer>, JpaSpecificationExecutor<CookiesConfigEntity> {
 
 	public List<CookiesConfigEntity> findAll();
 	
