@@ -2,17 +2,17 @@ package com.flower.spirit.dao;
 
 import java.util.List;
 
-import javax.transaction.Transactional;
+import jakarta.transaction.Transactional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 import com.flower.spirit.entity.DownloaderEntity;
 
 @Repository
 @Transactional
-public interface DownloaderDao extends PagingAndSortingRepository<DownloaderEntity, Integer>, JpaSpecificationExecutor<DownloaderEntity> {
+public interface DownloaderDao extends JpaRepository<DownloaderEntity, Integer>, JpaSpecificationExecutor<DownloaderEntity> {
 
 	DownloaderEntity findByStatus(String string);
 
