@@ -2,10 +2,10 @@ package com.flower.spirit.dao;
 
 import java.util.List;
 
-import javax.transaction.Transactional;
+import jakarta.transaction.Transactional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 import com.flower.spirit.entity.FfmpegQueueDataEntity;
@@ -14,7 +14,7 @@ import com.flower.spirit.entity.FfmpegQueueDataEntity;
 
 @Repository
 @Transactional
-public interface FfmpegQueueDataDao extends PagingAndSortingRepository<FfmpegQueueDataEntity, Integer>, JpaSpecificationExecutor<FfmpegQueueDataEntity> {
+public interface FfmpegQueueDataDao extends JpaRepository<FfmpegQueueDataEntity, Integer>, JpaSpecificationExecutor<FfmpegQueueDataEntity> {
 
 	List<FfmpegQueueDataEntity> findByStatus(String string);
 
