@@ -2,10 +2,10 @@ package com.flower.spirit.dao;
 
 import java.util.List;
 
-import javax.transaction.Transactional;
+import jakarta.transaction.Transactional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 import com.flower.spirit.entity.TikTokConfigEntity;
@@ -15,7 +15,7 @@ import com.flower.spirit.entity.TikTokConfigEntity;
 
 @Repository
 @Transactional
-public interface TikTokConfigDao extends PagingAndSortingRepository<TikTokConfigEntity, Integer>, JpaSpecificationExecutor<TikTokConfigEntity>{
+public interface TikTokConfigDao extends JpaRepository<TikTokConfigEntity, Integer>, JpaSpecificationExecutor<TikTokConfigEntity>{
 
 	
 	public List<TikTokConfigEntity> findAll();
