@@ -1,9 +1,9 @@
 package com.flower.spirit.dao;
 
-import javax.transaction.Transactional;
+import jakarta.transaction.Transactional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 import com.flower.spirit.entity.UserEntity;
@@ -11,7 +11,7 @@ import com.flower.spirit.entity.UserEntity;
 
 @Repository
 @Transactional
-public interface UserDao extends PagingAndSortingRepository<UserEntity, Integer>, JpaSpecificationExecutor<UserEntity>{
+public interface UserDao extends JpaRepository<UserEntity, Integer>, JpaSpecificationExecutor<UserEntity>{
 
 	UserEntity findByUsername(String username);
 	
