@@ -3,19 +3,18 @@ package com.flower.spirit.dao;
 import java.util.Date;
 import java.util.List;
 
-import javax.transaction.Transactional;
+import jakarta.transaction.Transactional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import com.flower.spirit.entity.VideoDataEntity;
-
-
 @Repository
 @Transactional
-public interface VideoDataDao extends PagingAndSortingRepository<VideoDataEntity, Integer>, JpaSpecificationExecutor<VideoDataEntity>{
+public interface VideoDataDao
+		extends JpaRepository<VideoDataEntity, Integer>, JpaSpecificationExecutor<VideoDataEntity> {
 
 	List<VideoDataEntity> findByVideoid(String videoid);
 
