@@ -2,10 +2,10 @@ package com.flower.spirit.dao;
 
 import java.util.List;
 
-import javax.transaction.Transactional;
+import jakarta.transaction.Transactional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 import com.flower.spirit.entity.VideoMixSegmentEntity;
 
@@ -14,7 +14,7 @@ import com.flower.spirit.entity.VideoMixSegmentEntity;
 
 @Repository
 @Transactional
-public interface VideoMixSegmentDao extends PagingAndSortingRepository<VideoMixSegmentEntity, Integer>, JpaSpecificationExecutor<VideoMixSegmentEntity>{
+public interface VideoMixSegmentDao extends JpaRepository<VideoMixSegmentEntity, Integer>, JpaSpecificationExecutor<VideoMixSegmentEntity>{
 
 	
 	public List<VideoMixSegmentEntity> findAll();
