@@ -4,7 +4,7 @@ import java.io.Serializable;
 import java.util.List;
 import java.util.Date;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 import com.flower.spirit.common.DataEntity;
 
