@@ -3,7 +3,7 @@ package com.flower.spirit.task;
 import java.util.List;
 import java.util.Set;
 
-import javax.annotation.PostConstruct;
+import jakarta.annotation.PostConstruct;
 
 import org.apache.commons.lang3.StringUtils;
 import org.quartz.CronExpression;
