@@ -1,20 +1,16 @@
 package com.flower.spirit.service;
 
-import java.io.BufferedReader;
 import java.io.File;
-import java.io.FileReader;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
+import java.util.Arrays;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.LinkedBlockingQueue;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import org.apache.commons.io.FilenameUtils;
@@ -23,7 +19,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
 import com.flower.spirit.config.Global;
 import com.flower.spirit.dao.VideoDataDao;
@@ -31,19 +26,15 @@ import com.flower.spirit.entity.ProcessHistoryEntity;
 import com.flower.spirit.entity.VideoDataEntity;
 import com.flower.spirit.utils.Aria2Util;
 import com.flower.spirit.utils.BiliUtil;
-import com.flower.spirit.utils.CommandUtil;
 import com.flower.spirit.utils.DateUtils;
 import com.flower.spirit.utils.DouUtil;
 import com.flower.spirit.utils.EmbyMetadataGenerator;
 import com.flower.spirit.utils.FileUtil;
-import com.flower.spirit.utils.FileUtils;
 import com.flower.spirit.utils.HttpUtil;
 import com.flower.spirit.utils.JsonChunkParser;
 import com.flower.spirit.utils.KuaishouParser;
 import com.flower.spirit.utils.KuaishouParser.VideoInfo;
-import com.flower.spirit.utils.Steamcmd;
 import com.flower.spirit.utils.StringUtil;
-import com.flower.spirit.utils.TikTokUtil;
 import com.flower.spirit.utils.URLUtil;
 import com.flower.spirit.utils.YtDlpUtil;
 import com.flower.spirit.utils.sendNotify;
